# docs/swagger.yaml
basePath: /api/v1
definitions:
  middleware.ErrorJsonResponse:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: Bad Request
        type: string
    type: object
...
  presenter.CustomerJsonResponse:
    properties:
      cpf:
        example: 123.456.789-00
        type: string
      created_at:
        example: "2024-02-09T10:00:00Z"
        type: string
      email:
        example: john.doe@email.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: John Doe
        type: string
      updated_at:
        example: "2024-02-09T10:00:00Z"
        type: string
    type: object
...
  presenter.OrderJsonPaginatedResponse:
    properties:
      limit:
        example: 10
        type: integer
      orders:
        items:
          $ref: '#/definitions/presenter.OrderJsonResponse'
        type: array
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
    type: object
  presenter.OrderJsonResponse:
    properties:
      created_at:
        example: "2024-02-09T10:00:00Z"
        type: string
      customer:
        $ref: '#/definitions/presenter.CustomerJsonResponse'
      customer_id:
        example: 1
        type: integer
      id:
        type: integer
      products:
        items:
          $ref: '#/definitions/presenter.ProductsJsonResponse'
        type: array
      status:
        example: PENDING
        type: string
      total_bill:
        example: "100.00"
        type: string
      updated_at:
        example: "2024-02-09T10:00:00Z"
        type: string
    type: object
...
  presenter.ProductsJsonResponse:
    properties:
      category_id:
        example: 1
        type: integer
      created_at:
        example: "2024-02-09T10:00:00Z"
        type: string
      description:
        example: Description of product A
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Product A
        type: string
      price:
        example: 99.99
        type: number
      quantity:
        type: integer
      updated_at:
        example: "2024-02-09T10:00:00Z"
        type: string
    type: object
...
  request.CreateOrderBodyRequest:
    properties:
      customer_id:
        example: 1
        type: integer
    required:
    - customer_id
    type: object
...
  request.UpdateOrderBodyRequest:
    properties:
      customer_id:
        example: 1
        type: integer
      staff_id:
        description: StaffID is only required when status is PREPARING, READY or COMPLETED
        example: 1
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/valueobject.OrderStatus'
        example: PENDING
    required:
    - customer_id
    - status
    type: object
  request.UpdateOrderPartilRequest:
    properties:
      staff_id:
        description: StaffID is only required when status is PREPARING, READY or COMPLETED
        example: 1
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/valueobject.OrderStatus'
        example: PENDING
    type: object
...
  valueobject.OrderStatus:
    enum:
    - OPEN
    - CANCELLED
    - PENDING
    - RECEIVED
    - PREPARING
    - READY
    - COMPLETED
    - UNDEFINDED
    type: string
    x-enum-varnames:
    - OPEN
    - CANCELLED
    - PENDING
    - RECEIVED
    - PREPARING
    - READY
    - COMPLETED
    - UNDEFINDED
...
info:
  contact: {}
  description: '### FIAP Tech Challenge Phase 3 - 10SOAT - G22'
  title: Fast Food API v3
  version: "1"
paths:
...
  /orders:
    get:
      consumes:
      - application/json
      description: |-
        List all orders
        ## Order list is sorted by:
        - **Status** in **descending** order (`READY` > `PREPARING` > `RECEIVED` > `PENDING` > `OPEN`)
        - **Created date** (CreatedAt) in **ascending** order (oldest first)
        Obs: Status CANCELLED and COMPLETED are not included in the list by default
      parameters:
      - description: Filter by customer ID
        in: query
        name: customer_id
        type: integer
      - description: 'Filter by status (Accept many), options: <sub>OPEN, PENDING,
          RECEIVED, PREPARING, READY</sub>, ex: <sub>PENDING</sub> or <sub>OPEN,PENDING</sub>'
        in: query
        name: status
        type: string
      - default: CANCELLED,COMPLETED
        description: 'Exclude by status (Accept many), options: <sub>NONE, OPEN, PENDING,
          RECEIVED, PREPARING, READY, CANCELLED, COMPLETED</sub>, ex: <sub>CANCELLED</sub>
          or <sub>CANCELLED,COMPLETED</sub> (default)'
        in: query
        name: status_exclude
        type: string
      - default: status:d,created_at
        description: Sort by field (Accept many). Use `<field_name>:d` for descending,
          and the default order is ascending
        in: query
        name: sort
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.OrderJsonPaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
      summary: List orders (Reference TC-1 2.b.vi; TC-2 1.a.iv)
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Creates a new order
      parameters:
      - description: Order data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/request.CreateOrderBodyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/presenter.OrderJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
      summary: Create order
      tags:
      - orders
  /orders/{id}:
    delete:
      description: Deletes a order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.OrderJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
      summary: Delete order
      tags:
      - orders
    get:
      consumes:
      - application/json
      description: Search for a order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.OrderJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
      summary: Get order
      tags:
      - orders
    patch:
      consumes:
      - application/json
      description: |-
        Partially updates an existing order
        The status are: **OPEN**, **CANCELLED**, **PENDING**, **RECEIVED**, **PREPARING**, **READY**, **COMPLETED**
        ## Transition of status:
        - OPEN      -> CANCELLED || PENDING
        - CANCELLED -> {},
        - PENDING   -> OPEN || RECEIVED
        - RECEIVED  -> PREPARING
        - PREPARING -> READY
        - READY     -> COMPLETED
        - COMPLETED -> {}
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Order data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/request.UpdateOrderPartilRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.OrderJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
      summary: Partial update order (Reference TC-2 1.a.v)
      tags:
      - orders
    put:
      consumes:
      - application/json
      description: |-
        Update an existing order
        The status are: **OPEN**, **CANCELLED**, **PENDING**, **RECEIVED**, **PREPARING**, **READY**, **COMPLETED**
        ## Transition of status:
        - OPEN      -> CANCELLED || PENDING
        - CANCELLED -> {},
        - PENDING   -> OPEN || RECEIVED
        - RECEIVED  -> PREPARING
        - PREPARING -> READY
        - READY     -> COMPLETED
        - COMPLETED -> {}
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Order data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/request.UpdateOrderBodyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.OrderJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorJsonResponse'
      summary: Update order
      tags:
      - orders
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...